renamed log parsers directory name to parsers
